# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Snake 3310 — API pública
  version: 0.0.0
  description: Saúde, prontidão, ranking e envio público (RN-0001, RN-0003, RN-0005 e RN-0006).
servers:
  - url: '{basePath}'
    variables:
      basePath:
        default: /snake-3310
        description: Prefixo BASE_PATH configurado no servidor; pode ser vazio na raiz.
paths:
  /api/health:
    get:
      operationId: checkLiveness
      summary: Processo HTTP está vivo, independente do banco (RN-0003)
      security: []
      responses:
        '200':
          description: Processo respondendo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
  /api/ready:
    get:
      operationId: checkReadiness
      summary: Banco responde à verificação (RN-0003)
      security: []
      responses:
        '200':
          description: Pronto para receber pedidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
        '503':
          description: Banco indisponível, sem mensagem interna.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/placares:
    get:
      operationId: listRanking
      summary: Lista os dez maiores placares públicos (RN-0001)
      description: Pontos em ordem decrescente; empates pelo envio mais antigo. Sem paginação nem parâmetros; campos de consulta desconhecidos recebem 400. Lista vazia quando não há placares.
      security: []
      responses:
        '200':
          description: Ranking limitado a dez entradas, sem datas nem identificadores internos.
          content:
            application/json:
              schema:
                type: object
                required: [scores]
                additionalProperties: false
                properties:
                  scores:
                    type: array
                    maxItems: 10
                    items:
                      type: object
                      required: [nickname, points]
                      additionalProperties: false
                      properties:
                        nickname:
                          type: string
                        points:
                          type: integer
                          minimum: 0
        '400':
          description: Parâmetro de consulta desconhecido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Falha interna ou de leitura do banco, sem detalhes internos.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      operationId: submitScore
      summary: Valida e registra um placar público (RN-0005)
      description: >-
        Somente nickname e points, sem coerção de tipos ou campos extras; consulta sem parâmetros.
        O servidor atribui a data UTC, filtra vocabulário ofensivo sem caixa/acentos e registra uma linha.
        Corpo máximo de 1024 bytes. Cinco tentativas por IP em uma janela fixa de 60 segundos, inclusive
        inválidas; contador local ao processo, até 4096 IPs, reiniciado com o processo. Saturação recusa
        novos IPs até expirar uma entrada. Somente o último proxy explicitamente confiado pode apresentar
        o cabeçalho saneado X-Snake-Client-IP; X-Forwarded-For nunca define a identidade.
        Não há autenticação nem prova criptográfica da partida; pontos plausíveis podem ser forjados.
        A interface bloqueia envios pendentes e novos envios após 201 na mesma partida. Não há chave de
        idempotência na API: cancelar a espera no navegador não desfaz uma gravação já concluída.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreSubmission'
      responses:
        '201':
          description: Placar gravado; data e identificador ficam no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreSubmission'
        '400':
          description: Corpo, consulta, apelido, pontos ou campos extras inválidos; nenhum placar gravado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '413':
          description: Corpo acima de 1024 bytes; nenhum placar gravado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '415':
          description: Tipo de conteúdo não suportado; nenhum placar gravado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Tentativas ou capacidade excedidas; nenhum placar gravado.
          headers:
            Retry-After:
              description: Segundos até uma nova tentativa, de 1 a 60.
              schema:
                type: integer
                minimum: 1
                maximum: 60
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Falha interna ou de gravação; resposta sem detalhes internos.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  schemas:
    ScoreSubmission:
      type: object
      required: [nickname, points]
      additionalProperties: false
      properties:
        nickname:
          type: string
          minLength: 3
          maxLength: 12
          pattern: '^[\p{L}\p{N}]{3,12}$'
          description: Letras ou números Unicode; termos do filtro local sem caixa/acentos são recusados.
        points:
          type: integer
          minimum: 0
          maximum: 1890
          multipleOf: 7
    Health:
      type: object
      required: [status]
      additionalProperties: false
      properties:
        status:
          type: string
          const: ok
    Problem:
      type: object
      required: [type, title, status]
      additionalProperties: false
      properties:
        type:
          type: string
          const: about:blank
        title:
          type: string
        status:
          type: integer
          enum: [400, 404, 413, 415, 429, 500, 503]